Feat: Add redesign of HomeActivity
Changes:
* Redesign BottomNavigationBar
* Add Outlined icons for BottomNavigationBar
* Redesign toolbar
* Set up global overflow menu
* Set up margins for toolbar
