+++ mifospay/src/main/res/drawable/bottomview_selector.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:color="@color/primaryDarkBlue" android:state_pressed="true"/>
-    <item android:color="@color/primaryDarkBlue" android:state_checked="true"/>
-    <item android:color="@color/grey_900"/>
+    <item android:color="@color/colorAccent" android:state_pressed="true"/>
+    <item android:color="@color/colorAccent" android:state_checked="true"/>
+    <item android:color="@color/colorAccentBlue50"/>
 </selector>
