+++ mifospay/src/main/res/drawable/ic_finance.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M21,7.28V5a2.006,2.006 0,0 0,-2 -2H5A2,2 0,0 0,3 5V19a2,2 0,0 0,2 2H19a2.006,2.006 0,0 0,2 -2V16.72A2,2 0,0 0,22 15V9a2,2 0,0 0,-1 -1.72ZM20,9v6H13V9ZM5,19V5H19V7H13a2.006,2.006 0,0 0,-2 2v6a2.006,2.006 0,0 0,2 2h6v2Z"
+      android:fillColor="#000000" />
+  <path
+      android:pathData="M16,12m-1.5,0a1.5,1.5 0,1 1,3 0a1.5,1.5 0,1 1,-3 0"
+      android:fillColor="#000000" />
+</vector>
