+++ mifospay/src/main/res/drawable/ic_home.xml
@@ -1,9 +1,9 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-        android:width="24dp"
-        android:height="24dp"
-        android:viewportHeight="24.0"
-        android:viewportWidth="24.0">
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M10,20v-6h4v6h5v-8h3L12,3 2,12h3v8z"/>
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M12,5.69l5,4.5V18H15V12H9v6H7V10.19l5,-4.5M12,3 L2,12H5v8h6V14h2v6h6V12h3Z"
+      android:fillColor="#000000"/>
 </vector>
