+++ mifospay/src/main/res/drawable/ic_person.xml
@@ -1,9 +1,9 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-        android:width="24dp"
-        android:height="24dp"
-        android:viewportHeight="24.0"
-        android:viewportWidth="24.0">
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M12,12c2.21,0 4,-1.79 4,-4s-1.79,-4 -4,-4 -4,1.79 -4,4 1.79,4 4,4zM12,14c-2.67,0 -8,1.34 -8,4v2h16v-2c0,-2.66 -5.33,-4 -8,-4z"/>
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M12,2A10,10 0,1 0,22 12,10 10,0 0,0 12,2ZM7.07,18.28c0.43,-0.9 3.05,-1.78 4.93,-1.78s4.51,0.88 4.93,1.78a7.925,7.925 0,0 1,-9.86 0ZM18.36,16.83c-1.43,-1.74 -4.9,-2.33 -6.36,-2.33s-4.93,0.59 -6.36,2.33a8,8 0,1 1,12.72 0ZM12,6a3.5,3.5 0,1 0,3.5 3.5A3.491,3.491 0,0 0,12 6ZM12,11a1.5,1.5 0,1 1,1.5 -1.5A1.5,1.5 0,0 1,12 11Z"
+      android:fillColor="#000000" />
 </vector>
