+++ mifospay/src/main/res/menu/menu_overflow.xml
@@ -1,12 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
-      xmlns:app="http://schemas.android.com/apk/res-auto"
-      xmlns:tools="http://schemas.android.com/tools"
-      tools:context=".home.ui.HomeActivity">
-
-    <item
-        android:id="@+id/item_edit_profile"
-        android:title="Edit Profile"/>
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".home.ui.HomeActivity">
 
     <item
         android:id="@+id/item_faq"
