+++ mifospay/src/main/res/values/preloaded_fonts.xml
@@ -2,5 +2,6 @@
 <resources>
     <array name="preloaded_fonts" translatable="false">
         <item>@font/roboto</item>
+        <item>@font/roboto_medium</item>
     </array>
 </resources>
